requirements.txt
fastapi
uvicorn
pandas
numpy
torch
transformers
pinecone-client
groq
pydantic